fix bug

+++ src/page/SignTracking.jsx
@@ -44,7 +44,7 @@ const SignTracking = () => {
       // 4. TODO - Make Detections
       const img = tf.browser.fromPixels(video);
       const resized = tf.image.resizeBilinear(img, [640, 480]);
-      const casted = resized.cast("int32");
+      const casted = resized.cast("float32");
       const expanded = casted.expandDims(0);
       const obj = await net.executeAsync(expanded);
       console.log(obj);
